Crate nmap_service_probes

nmap-service-probes

A #[no_std] parser and code generator for the nmap-service-probes file format.

let probes = nmap_service_probes::parse("Probe UDP Help q|help\r\n\r\n|")?;

assert_eq!(
    &probes[0],
    &ServiceProbe {
        protocol: Protocol::Udp,
        name: "Help".into(),
        string: "help\r\n\r\n".into(),
        no_payload: false,
        total_wait_millis: None,
        fallbacks: <_>::default(),
        ports: vec![],
        rarity: None,
        ssl_ports: <_>::default(),
        tcp_wrapped_millis: <_>::default(),
        matches: vec![]
    }
);

Structs

Match

Represents how to recognize services based on responses to the string sent by the probe.

NmapServiceProbes

A structured representation of a nmap-service-probes document.

Regex

A regular expression. These patterns are “Perl” or PCRE regular expressions.

ServiceProbe

Represents the data needed to define a probe for a given service.

Enums

Error

Potential errors that can occur when constructing NmapServiceProbes from a nmap-service-probes document.

MatchKind

The kind of match behaviour that should be used.

Port

The port number or range of numbers.

PortProtocol

Represents the set of ports and protocols used for exclusion.

Protocol

The transport protocol of the service.

Rarity

Corresponds to how rare a given service is considered, and thus how frequently the probe is expected to run. The higher the number, the rarer.

Functions

parse

Accepts a nmap-service-probes document, returning NmapServiceProbes if successful transforming the line based document format into a set of datatypes with associated fields.